+++ docs/CHANGELOG.md
@@ -7,6 +7,10 @@ Version 3.45.0 (July 2, 2024)
 
 Added a `Tree` argument to `AnnotatedTypes.adaptParameters()`
 
+Deprecated methods:
+ * `TreeUtils.isVarArgs()` => `isVarargsCall()`
+ * `TreeUtils.isVarArgMethodCall()` => `isVarargsCall()`
+
 **Closed issues:**